PayMe/apps/api/pyproject.toml
[project]
name = "payme-lite-api"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "alembic>=1.14.0",
  "argon2-cffi>=23.1.0",
  "cryptography>=42.0.0",
  "fastapi>=0.115.0",
  "email-validator>=2.2.0",
  "numpy>=1.26.0",
  "passlib[bcrypt]>=1.7.4",
  "psycopg[binary]>=3.2.0",
  "pydantic-settings>=2.5.0",
  "python-jose[cryptography]>=3.3.0",
  "scipy>=1.13.0",
  "sqlalchemy>=2.0.35",
  "structlog>=24.4.0",
  "uvicorn>=0.30.0"
]

[project.optional-dependencies]
dev = [
  "black>=24.8.0",
  "httpx>=0.27.0",
  "pytest>=8.3.0",
  "pytest-cov>=5.0.0",
  "ruff>=0.6.0"
]
gmail = [
  "google-auth-oauthlib>=1.2.0",
  "google-api-python-client>=2.140.0"
]
plaid = [
  "plaid-python>=20.0.0"
]

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["app/tests"]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
